feat: mvp de combate, data.sql limpio y fixes de compilacion

+++ src/main/java/com/tallerwebi/dominio/ServicioCombate.java
@@ -1,7 +1,7 @@
 package com.tallerwebi.dominio;
 
 public interface ServicioCombate {
-  Integer jugarCarta(Integer identificadorCarta, Long identificadorPartida);
+  Integer jugarCarta(Long identificadorCarta, Long identificadorPartida);
 
   Partida obtenerPartidaPorIdentificador(Long identificadorPartida);
 }
